Fixed up several warning types: const, casting, incorrect sizes, signed/unsigned comparisons.

+++ dpkg-deb/extract.c
@@ -85,7 +85,7 @@ void extracthalf(const char *debar, const char *directory,
   char versionbuf[40];
   float versionnum;
   char ctrllenbuf[40], *infobuf;
-  long ctrllennum, memberlen= 0;
+  size_t ctrllennum, memberlen= 0;
   int dummy, l= 0;
   pid_t c1=0,c2,c3;
   unsigned char *ctrlarea= 0;
